books_data/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
tqdm==4.66.1
python-dateutil==2.8.2 